Extract index annotation building logic
